backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
djongo==1.3.6
pymongo==3.12.3
mongoengine==0.27.0
Pillow==10.1.0
django-storages==1.14.2
boto3==1.33.0
celery==5.3.4
redis==5.0.1
django-filter==23.3
djangorestframework-filters==1.0.0.dev2
drf-spectacular==0.26.5
django-ratelimit==4.1.0
python-decouple==3.8
django-environ==0.11.2
qrcode==7.4.2
Pillow==10.1.0
geopy==2.4.1
haversine==2.8.0
django-celery-beat==2.5.0
flower==2.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
factory-boy==3.3.0
pytest-django==4.7.0
coverage==7.3.2
